Rewrote the program to utilize the digitalWriteByte() function that should increase speed by up to 8x

+++ Project Files/cPi/cprogram/sendInstruction.h
@@ -1,6 +1,6 @@
 #ifndef SENDINSTRUCTION_DOT_H
 #define SENDINSTRUCTION_DOT_H
 
-int sendInstruction(uint8_t msB, long long instruction);
+int sendInstruction(int msB, long long instruction);
 
 #endif
